Add new container option and lint
