+++ frontend/pages/index.js
@@ -1,7 +1,11 @@
+import React from 'react';
+
+import Container from 'react-bootstrap/Container';
+
 export default props => {
   return (
-    <>
+    <Container fluid>
       <h1>Welcome!</h1>
-    </>
+    </Container>
   );
 };
