+++ frontend/src/components/FullWidthContainer.scss
@@ -1,0 +1,4 @@
+.full-width-container {
+    padding-left: 0;
+    padding-right: 0;
+}
